fix : move main function to cli.py
