+++ zenix/functions.py
@@ -153,43 +153,3 @@ def write_wav(filepath: str, audio: np.ndarray, sample_rate: int) -> None:
 
 
 
-def main() -> None:
-    """
-    CLI entry point.
-    """
-    args = parse_args()
-
-    if not 0.0 <= args.volume <= 1.0:
-        print("Volume must be between 0 and 1.")
-        sys.exit(1)
-
-    print(f"🔊 Playing {args.type} noise... Press Ctrl+C to stop.")
-
-    def handler(sig, frame) -> None:
-        print("\nStopping...")
-        sys.exit(0)
-
-    signal.signal(signal.SIGINT, handler)
-
-    audio = generate_noise(
-        noise_type=args.type,
-        duration=args.duration,
-        sample_rate=DEFAULT_SAMPLE_RATE,
-        volume=args.volume,
-        fade_in=args.fade_in,
-    )
-
-    with tempfile.NamedTemporaryFile(suffix=".wav", delete=False) as tmp:
-        temp_path = tmp.name
-
-    try:
-        write_wav(temp_path, audio, DEFAULT_SAMPLE_RATE)
-
-        if args.loop:
-            play(temp_path, loop=True, async_mode=True)
-            signal.pause()
-        else:
-            play(temp_path)
-    finally:
-        if os.path.exists(temp_path):
-            os.remove(temp_path)
